+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Version [4.0.0](https://github.com/cedx/yii2-json-messages/compare/v3.0.0...v4.0.0)
 - Breaking change: raised the required [PHP](https://secure.php.net) version.
+- Breaking change: removed the `jsonSerialize()` and `__toString()` methods.
 - Breaking change: using PHP 7.1 features, like void functions.
 
 ## Version [3.0.0](https://github.com/cedx/yii2-json-messages/compare/v2.4.0...v3.0.0)